refactor(trace): lift run-timeline row fetch into the runtime crate
