+++ crates/defra-agent-cli/src/commands/trace.rs
@@ -16,6 +16,7 @@ use defra_agent::run_timeline::{
     TimelineInferenceCallRow, TimelineMessageRow, TimelineRequestEvent, TimelineRequestRow,
     TimelineResponseRow, TimelineSessionRow, TimelineToolCallEvent, TimelineToolCallRow,
 };
+use defra_agent::run_timeline_fetch::{load_run_timeline, load_run_timeline_rows};
 use defra_agent::trace_export::{
     analyze_request_failure, analyze_tool_call, extract_raw_tool_call_json, latency_ms,
     raw_message_json, AmyToolCallTraceRecord,
@@ -120,67 +121,6 @@ fn trace_project_schema(args: TraceProjectSchemaArgs) -> Result<()> {
         print_json(&schema)?;
     }
     Ok(())
-}
-
-async fn load_run_timeline(access: &ConfigAccess, request_id: &str) -> Result<RunTimeline> {
-    Ok(build_run_timeline(
-        load_run_timeline_rows(access, request_id).await?,
-    ))
-}
-
-async fn load_run_timeline_rows(
-    access: &ConfigAccess,
-    request_id: &str,
-) -> Result<RunTimelineRows> {
-    let request = load_timeline_request_by_id(access, request_id).await?;
-    let root_session_id = request.session_id.clone();
-
-    let mut requests = match root_session_id.as_deref() {
-        Some(session_id) => load_timeline_requests_for_session(access, session_id).await?,
-        None => Vec::new(),
-    };
-    merge_timeline_request(&mut requests, request.clone());
-    for child in load_timeline_child_requests(access, &request.request_id).await? {
-        merge_timeline_request(&mut requests, child);
-    }
-
-    let session_ids = timeline_session_ids(&requests);
-    let mut messages = Vec::new();
-    let mut tool_calls = Vec::new();
-    let mut responses = Vec::new();
-    for session_id in &session_ids {
-        messages.extend(load_timeline_messages_for_session(access, session_id).await?);
-        tool_calls.extend(load_timeline_tool_calls_for_session(access, session_id).await?);
-        responses.extend(load_timeline_responses_for_session(access, session_id).await?);
-    }
-    if session_ids.is_empty() || root_session_id.is_none() {
-        responses.extend(load_timeline_responses_for_request(access, &request.request_id).await?);
-    }
-    let mut inference_calls = Vec::new();
-    for request_id in timeline_request_ids(&requests) {
-        inference_calls
-            .extend(load_timeline_inference_calls_for_request(access, &request_id).await?);
-    }
-
-    let session = match root_session_id.as_deref() {
-        Some(session_id) => load_timeline_session(access, session_id).await?,
-        None => None,
-    };
-    let conversation = match root_session_id.as_deref() {
-        Some(session_id) => load_timeline_conversation(access, session_id).await?,
-        None => None,
-    };
-
-    Ok(RunTimelineRows {
-        request,
-        session,
-        conversation,
-        requests,
-        messages,
-        tool_calls,
-        inference_calls,
-        responses,
-    })
 }
 
 #[derive(Debug, Clone)]
@@ -1151,395 +1091,6 @@ async fn load_request_by_id(access: &ConfigAccess, request_id: &str) -> Result<R
         .into_iter()
         .next()
         .ok_or_else(|| anyhow::anyhow!("request {request_id} not found"))
-}
-
-async fn load_timeline_request_by_id(
-    access: &ConfigAccess,
-    request_id: &str,
-) -> Result<TimelineRequestRow> {
-    let query = format!(
-        r#"{{
-            AgentRequest(
-                filter: {{ request_id: {{ _eq: "{}" }} }},
-                order: {{ created_at: DESC }},
-                limit: 1
-            ) {{
-                _docID
-                request_id
-                agent_did
-                behavior_id
-                session_id
-                content
-                metadata
-                status
-                lifecycle_state
-                backend_id
-                failure_reason
-                created_at
-                retry_count
-                interrupt_requested_at
-                caused_by_parent_request_id
-                caused_by_parent_tool_call_id
-            }}
-        }}"#,
-        escape_graphql_string(request_id)
-    );
-    load_rows::<TimelineRequestRow>(access, "AgentRequest", &query)
-        .await?
-        .into_iter()
-        .next()
-        .ok_or_else(|| anyhow::anyhow!("request {request_id} not found"))
-}
-
-async fn load_timeline_requests_for_session(
-    access: &ConfigAccess,
-    session_id: &str,
-) -> Result<Vec<TimelineRequestRow>> {
-    let query = format!(
-        r#"{{
-            AgentRequest(
-                filter: {{ session_id: {{ _eq: "{}" }} }},
-                order: {{ created_at: ASC }}
-            ) {{
-                _docID
-                request_id
-                agent_did
-                behavior_id
-                session_id
-                content
-                metadata
-                status
-                lifecycle_state
-                backend_id
-                failure_reason
-                created_at
-                retry_count
-                interrupt_requested_at
-                caused_by_parent_request_id
-                caused_by_parent_tool_call_id
-            }}
-        }}"#,
-        escape_graphql_string(session_id)
-    );
-    load_rows(access, "AgentRequest", &query).await
-}
-
-async fn load_timeline_child_requests(
-    access: &ConfigAccess,
-    parent_request_id: &str,
-) -> Result<Vec<TimelineRequestRow>> {
-    let query = format!(
-        r#"{{
-            AgentRequest(
-                filter: {{ caused_by_parent_request_id: {{ _eq: "{}" }} }},
-                order: {{ created_at: ASC }}
-            ) {{
-                _docID
-                request_id
-                agent_did
-                behavior_id
-                session_id
-                content
-                metadata
-                status
-                lifecycle_state
-                backend_id
-                failure_reason
-                created_at
-                retry_count
-                interrupt_requested_at
-                caused_by_parent_request_id
-                caused_by_parent_tool_call_id
-            }}
-        }}"#,
-        escape_graphql_string(parent_request_id)
-    );
-    load_rows(access, "AgentRequest", &query).await
-}
-
-async fn load_timeline_messages_for_session(
-    access: &ConfigAccess,
-    session_id: &str,
-) -> Result<Vec<TimelineMessageRow>> {
-    let query = format!(
-        r#"{{
-            AgentMessage(
-                filter: {{ session_id: {{ _eq: "{}" }} }},
-                order: {{ sequence: ASC }}
-            ) {{
-                _docID
-                session_id
-                request_id
-                sequence
-                role
-                content
-                timestamp
-            }}
-        }}"#,
-        escape_graphql_string(session_id)
-    );
-    match load_rows(access, "AgentMessage", &query).await {
-        Ok(rows) => Ok(rows),
-        Err(error) if error.to_string().contains("request_id") => {
-            let fallback_query = format!(
-                r#"{{
-                    AgentMessage(
-                        filter: {{ session_id: {{ _eq: "{}" }} }},
-                        order: {{ sequence: ASC }}
-                    ) {{
-                        _docID
-                        session_id
-                        sequence
-                        role
-                        content
-                        timestamp
-                    }}
-                }}"#,
-                escape_graphql_string(session_id)
-            );
-            load_rows(access, "AgentMessage", &fallback_query).await
-        }
-        Err(error) => Err(error),
-    }
-}
-
-async fn load_timeline_tool_calls_for_session(
-    access: &ConfigAccess,
-    session_id: &str,
-) -> Result<Vec<TimelineToolCallRow>> {
-    let query = format!(
-        r#"{{
-            AgentToolCall(
-                filter: {{ session_id: {{ _eq: "{}" }} }},
-                order: {{ started_at: ASC }}
-            ) {{
-                _docID
-                request_id
-                session_id
-                message_sequence
-                tool_name
-                tool_call_id
-                args
-                result
-                status
-                lifecycle_state
-                started_at
-                deadline_at
-                completed_at
-                selected_service_id
-                selected_tool_name
-                tool_failure_class
-                denial_reason
-                denied_argv
-                denied_command
-                denied_argument
-                denied_subcommand
-                denied_prefix
-                policy_mode
-                policy_network
-                latency_ms
-                await_mode
-                cancel_policy
-                cancel_cause
-                child_request_id
-            }}
-        }}"#,
-        escape_graphql_string(session_id)
-    );
-    load_rows(access, "AgentToolCall", &query).await
-}
-
-async fn load_timeline_responses_for_session(
-    access: &ConfigAccess,
-    session_id: &str,
-) -> Result<Vec<TimelineResponseRow>> {
-    let query = format!(
-        r#"{{
-            AgentResponse(
-                filter: {{ session_id: {{ _eq: "{}" }} }},
-                order: {{ created_at: ASC }}
-            ) {{
-                _docID
-                request_id
-                agent_did
-                behavior_id
-                session_id
-                content
-                reasoning
-                status
-                error_message
-                token_count
-                progress_seq
-                materialized_message_sequence
-                materialized_at
-                created_at
-                completed_at
-                interrupted_at
-            }}
-        }}"#,
-        escape_graphql_string(session_id)
-    );
-    load_rows(access, "AgentResponse", &query).await
-}
-
-async fn load_timeline_responses_for_request(
-    access: &ConfigAccess,
-    request_id: &str,
-) -> Result<Vec<TimelineResponseRow>> {
-    let query = format!(
-        r#"{{
-            AgentResponse(
-                filter: {{ request_id: {{ _eq: "{}" }} }},
-                order: {{ created_at: ASC }}
-            ) {{
-                _docID
-                request_id
-                agent_did
-                behavior_id
-                session_id
-                content
-                reasoning
-                status
-                error_message
-                token_count
-                progress_seq
-                materialized_message_sequence
-                materialized_at
-                created_at
-                completed_at
-                interrupted_at
-            }}
-        }}"#,
-        escape_graphql_string(request_id)
-    );
-    load_rows(access, "AgentResponse", &query).await
-}
-
-async fn load_timeline_inference_calls_for_request(
-    access: &ConfigAccess,
-    request_id: &str,
-) -> Result<Vec<TimelineInferenceCallRow>> {
-    let query = format!(
-        r#"{{
-            InferenceCall(
-                filter: {{ request_id: {{ _eq: "{}" }} }},
-                order: {{ call_seq: ASC }}
-            ) {{
-                _docID
-                call_id
-                request_id
-                call_seq
-                attempt
-                call_state
-                failure_reason
-                queued_at
-                started_at
-                ended_at
-                backend_id
-                call_kind
-            }}
-        }}"#,
-        escape_graphql_string(request_id)
-    );
-    load_rows(access, "InferenceCall", &query).await
-}
-
-async fn load_timeline_session(
-    access: &ConfigAccess,
-    session_id: &str,
-) -> Result<Option<TimelineSessionRow>> {
-    let query = format!(
-        r#"{{
-            AgentSession(
-                filter: {{ session_id: {{ _eq: "{}" }} }},
-                limit: 1
-            ) {{
-                _docID
-                session_id
-                agent_name
-                behavior_id
-                started
-                ended
-                status
-            }}
-        }}"#,
-        escape_graphql_string(session_id)
-    );
-    Ok(
-        load_rows::<TimelineSessionRow>(access, "AgentSession", &query)
-            .await?
-            .into_iter()
-            .next(),
-    )
-}
-
-async fn load_timeline_conversation(
-    access: &ConfigAccess,
-    session_id: &str,
-) -> Result<Option<TimelineConversationRow>> {
-    let query = format!(
-        r#"{{
-            AgentConversation(
-                filter: {{ session_id: {{ _eq: "{}" }} }},
-                limit: 1
-            ) {{
-                _docID
-                session_id
-                agent_name
-                agent_did
-                behavior_id
-                title
-                title_source
-                preview_text
-                status
-                created_at
-                updated_at
-                latest_request_id
-                forked_from_session_id
-            }}
-        }}"#,
-        escape_graphql_string(session_id)
-    );
-    Ok(
-        load_rows::<TimelineConversationRow>(access, "AgentConversation", &query)
-            .await?
-            .into_iter()
-            .next(),
-    )
-}
-
-fn merge_timeline_request(rows: &mut Vec<TimelineRequestRow>, request: TimelineRequestRow) {
-    if !rows.iter().any(|row| row.request_id == request.request_id) {
-        rows.push(request);
-    }
-}
-
-fn timeline_session_ids(requests: &[TimelineRequestRow]) -> Vec<String> {
-    requests
-        .iter()
-        .filter_map(|request| {
-            request
-                .session_id
-                .as_deref()
-                .map(str::trim)
-                .filter(|value| !value.is_empty())
-                .map(ToOwned::to_owned)
-        })
-        .collect::<BTreeSet<_>>()
-        .into_iter()
-        .collect()
-}
-
-fn timeline_request_ids(requests: &[TimelineRequestRow]) -> Vec<String> {
-    requests
-        .iter()
-        .filter_map(|request| {
-            let request_id = request.request_id.trim();
-            (!request_id.is_empty()).then_some(request_id.to_string())
-        })
-        .collect::<BTreeSet<_>>()
-        .into_iter()
-        .collect()
 }
 
 async fn load_requests_for_sessions(
